Added new algorithms: find_match, starts_with, ends_with, first_result (_if, _opt), and iota.
See: https://original.fileswhatever.net/algorithms.php

Added:
- transform_exclusive_scan
- exchange
- type traits: is_input_iterator, is_forward_iterator,
is_bidirectional_iterator, is_random_access_iterator
- detail_sort::merge_sort, detail_sort::heap_sort,
detail_sort::counting_sort

Changed:
- Renamed is_output_iterator to is_output_iterator_for.

Fixed:
- is_iterator did not match the definition in the standard

Improved:
- contains: rebased on any_of, added iterator overload, improved
constraints
- find_in*: explicitly cast to unsigned
- various algorithms: changed all calls to parameter operations to use
invoke
- nl: use correct EOF
- unformatted_expect: correctly check for EOF
- incrementer, decrementer: added function call operators
- magic_enumerate: removed volatile from index type

Refactoring:
- Made several integral conversions explicit, or removed them entirely
when possible. direct_map and range_t cannot be fixed without major
design changes.
- Made FP-to-integral casts in count_digits explicit.
- Fixed detail_direct_map::range_of for signed types and partially for
large types.
- Removed shadowing of nums::min and nums::max
- Improved some of the tests

Began work on adjacent_iterator
Re-enabled several warning flags

This commit was much too large, but I was mixing coding with other work
and ended up with a constantly shifting list of partial features

+++ kblib/format.h
@@ -71,7 +71,8 @@ constexpr auto count_digits(Number val)
 	if (val == 0 or val == 1) {
 		return 1;
 	} else {
-		return std::ceil(std::nextafter(std::log10(std::fabs(val)), INFINITY)) +
+		return static_cast<int>(std::ceil(
+		           std::nextafter(std::log10(std::fabs(val)), INFINITY))) +
 		       (val < 0);
 	}
 }
@@ -88,9 +89,9 @@ constexpr auto count_digits(Number val)
 	if (val == 0) {
 		return 1;
 	} else if (val == static_cast<Number>(-1)) {
-		return std::ceil(std::log10(val));
+		return static_cast<int>(std::ceil(std::log10(val)));
 	} else {
-		return std::ceil(std::log10(val + 1));
+		return static_cast<int>(std::ceil(std::log10(val + 1)));
 	}
 }
 
@@ -107,9 +108,11 @@ constexpr auto count_digits(Number val, int base) -> int {
 	if (val == 0) {
 		return 1;
 	} else if (std::is_floating_point<Number>::value) {
-		return std::ceil(std::numeric_limits<Number>::digits * std::logb(base));
+		return static_cast<int>(
+		    std::ceil(std::numeric_limits<Number>::digits * std::logb(base)));
 	} else {
-		return std::ceil(std::log(std::abs(val) + 1) / std::log(base)) +
+		return static_cast<int>(
+		           std::ceil(std::log(std::abs(val) + 1) / std::log(base))) +
 		       (val < 0);
 	}
 }
